refactor: port snackbar to typescript
